Create Rotate Array
The function takes an array (nums) and the number of steps to rotate (k). The effective rotation steps are calculated by taking the modulo with the length of the array to handle cases where k is greater than the length of the array.

The rotation is performed in three steps:

Reverse the entire array.
Reverse the first k elements.
Reverse the remaining elements.