Linked list Question 2
To find the node where the cycle begins in a linked list, you can modify Floyd's cycle-finding algorithm. After detecting the cycle, you can use another pointer to find the start of the cycle. This pointer starts from the head of the list and moves together with the slow pointer until they meet. The meeting point will be the start of the cycle.